warning fix

+++ src/libraries/simulation_kernel/Entities/Agents/Roles/Posture/PHY_RolePion_Posture.cpp
@@ -507,7 +507,7 @@ void PHY_RolePion_Posture::Execute( detection::PerceptionDistanceComputer_ABC& a
 void PHY_RolePion_Posture::Execute( urbanLocation::UrbanLocationComputer_ABC& algorithm ) const
 {
     if( pCurrentPosture_ == &PHY_Posture::poste_ )
-        algorithm.SetUrbanDeployment( rPostureCompletionPercentage_ );
+        algorithm.SetUrbanDeployment( static_cast< float >( rPostureCompletionPercentage_ ) );
     if( pCurrentPosture_ == &PHY_Posture::posteAmenage_ )
         algorithm.SetUrbanDeployment( 1.f );
 }
